Add new pr-commenter custom task
Part of #483

This custom task is meant to be used alongside PR status report updates to add/update/delete a comment equivalent to the Prow one listing job failures with links to logs and the command to re-test.

I initially made an attempt to update the `github-add-comment` task in the catalog to do this (https://github.com/tektoncd/catalog/pull/1051) but decided that a custom task frankly made a lot more sense. Its behavior is a pretty direct emulation of how Prow and Lighthouse do this, looking for existing comments by the bot user with a given tag in the comment, and updating/deleting that comment if it exists, creating a new comment otherwise. It's a bit simplified, and instead of using `ProwJob`s to know what should still be in the comment and what should be removed, it specifically just adds or removes the specific job it's reporting on while preserving whatever else was there.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>